% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/utils.R
\name{add_list_to_list}
\alias{add_list_to_list}
\title{Add the elements of the second list to the first and return the result.}
\usage{
add_list_to_list(l1, l2, ...)
}
\arguments{
\item{l1}{the list to which entries should get added}

\item{l2}{the list from which entries should get added}

\item{...}{names of entries to add from list l2 to l1}
}
\value{
a copy of l1 with the entries from l2, which are not already present, added.
(if names were also specified, only entries which correspond to one of those
names and are not already present)
}
\description{
Only elements not already there are added, and if additional parameters
are given, they are interpreted as element names and only those names
are added from the second list.
}

