% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{bootPRF}
\alias{bootPRF}
\title{Return a list of bootstrap objects for the precision and recall values}
\usage{
bootPRF(data)
}
\arguments{
\item{data}{expected to be a data frame as stored within an evaluation object.}
}
\description{
If the data dataframe does not contain the necessary rows for individual
documents, an error is raised.
}

