% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/load.R
\name{GateEval}
\alias{GateEval}
\title{Read a file as created by one of the Evaluation PRs}
\usage{
GateEval(filename, evalId = NULL)
}
\arguments{
\item{filename}{The name of the file to read.}

\item{evalId}{The evaluation id to select, if the file contains rows for more than one evaluation id.
If this is not specified, all evaluation ids will be loaded.}
}
\value{
an evaluation container object. The container object allows to access individual evaluation objects
  using the \code{get} function.
}
\description{
Read a file as created by one of the Evaluation PRs
}
\examples{
\dontrun{
  GateEval("thisid.tsv", evalId = "thisid")
}
}

