% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/evallist.R
\name{GateEvalList}
\alias{GateEvalList}
\title{Read in a list of evaluation files and select a specific evaluation type, id or annotation type.}
\usage{
GateEvalList(fileNames, evaluationType = NULL, evaluationId = NULL,
  annotationType = NULL)
}
\arguments{
\item{fileNames}{a vector of file names to read in}
}
\value{
an object that contains the list of evaluations
}
\description{
NOTE: the evaluation instances are stored by a name "eval<N>" where "N" is the
index of the file.
}

