% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/GateEvalR.R
\docType{package}
\name{GateEvalR}
\alias{GateEvalR}
\alias{GateEvalR-package}
\title{GateEvalR: A Package to Help Analysing GATE NLP Evaluation Output Files.}
\description{
The GateEvalR package can read in the TSV (tab-separated values) files created
by the evaluation PRs from the \code{Evaluation} plugin for for the
GATE NLP framework (\url{https://gate.ac.uk}). The files are converted into
R objects of verious types, depending on the type of evaluation that was
performed. These objects can be plotted, summarized or viewed using the standard
R functions and there are additional functions for specific kinds of analyses
or the calculation of specific statistics.
}
\section{New Functions}{

}

