% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/utils.R
\name{getAt}
\alias{getAt}
\title{Same as [] but if the index is NA, returns NA as the element}
\usage{
getAt(vector, index)
}
\description{
Same as [] but if the index is NA, returns NA as the element
}

