% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/load.R
\name{initializeObject}
\alias{initializeObject}
\title{Internal generic function for initializing an evaluation instance.}
\usage{
initializeObject(x, ...)
}
\arguments{
\item{x}{the object to initialize}

\item{...}{additional parameters}
}
\value{
the initialized object
}
\description{
Internal generic function for initializing an evaluation instance.
}

