% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/taggingbasic.R
\name{initializeObject.TaggingBasic}
\alias{initializeObject.TaggingBasic}
\title{Initialize an object of type Tagging Basic}
\usage{
\method{initializeObject}{TaggingBasic}(x)
}
\arguments{
\item{x}{the object to initialize}
}
\value{
the initialized object
}
\description{
Initialize an object of type Tagging Basic
}

