% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/taggingdetaillistmatches.R
\name{initializeObject.TaggingDetailListMatches}
\alias{initializeObject.TaggingDetailListMatches}
\title{Initialize an object of type TaggingDetailListMatches}
\usage{
\method{initializeObject}{TaggingDetailListMatches}(x)
}
\arguments{
\item{x}{the object to initialize}
}
\value{
the initialized object
}
\description{
Initialize an object of type TaggingDetailListMatches
}

