% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/tagginglistdisamb.R
\name{initializeObject.TaggingListDisamb}
\alias{initializeObject.TaggingListDisamb}
\title{Initialize an object of type Tagging List Disamb}
\usage{
\method{initializeObject}{TaggingListDisamb}(x)
}
\arguments{
\item{x}{the object to initialize}
}
\value{
the initialized object
}
\description{
Initialize an object of type Tagging List Disamb
}

