% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/tagginglistdisambbest.R
\name{initializeObject.TaggingListDisambBest}
\alias{initializeObject.TaggingListDisambBest}
\title{Initialize an object of type TaggingListBest}
\usage{
\method{initializeObject}{TaggingListDisambBest}(x)
}
\arguments{
\item{x}{the object to initialize}
}
\value{
the initialized object
}
\description{
Initialize an object of type TaggingListBest
}

