% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/tagginglistscore.R
\name{initializeObject.TaggingListScore}
\alias{initializeObject.TaggingListScore}
\title{Initialize an object of type Tagging List Score}
\usage{
\method{initializeObject}{TaggingListScore}(x)
}
\arguments{
\item{x}{the object to initialize}
}
\value{
the initialized object
}
\description{
Initialize an object of type Tagging List Score
}

