% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/taggingscore.R
\name{initializeObject.TaggingScore}
\alias{initializeObject.TaggingScore}
\title{Initialize an object of type Tagging Score}
\usage{
\method{initializeObject}{TaggingScore}(x)
}
\arguments{
\item{x}{the object to initialize}
}
\value{
the initialized object
}
\description{
Initialize an object of type Tagging Score
}

