% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/taggingbasic.R
\name{plot.TaggingBasic}
\alias{plot.TaggingBasic}
\title{Plot an object of type TaggingBasic}
\usage{
\method{plot}{TaggingBasic}(x, strict = TRUE, addstripchart = FALSE, ...)
}
\description{
Plot an object of type TaggingBasic
}

