% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/taggingdetaillistmatches.R
\name{plot.TaggingDetailListMatches}
\alias{plot.TaggingDetailListMatches}
\title{Plot an object of type TaggingListBest}
\usage{
\method{plot}{TaggingDetailListMatches}(x, strict = TRUE, maxRank = NA,
  minRank = NA, ...)
}
\description{
By default this shows the distribution of the rank of the first strict
match, for all rows which do have a strict match
}

