% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/tagginglistbest.R
\name{plot.TaggingListBest}
\alias{plot.TaggingListBest}
\title{Plot an object of type TaggingListBest}
\usage{
\method{plot}{TaggingListBest}(x, strict = TRUE, addstripchart = FALSE, ...)
}
\description{
Plot an object of type TaggingListBest
}

