% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/tagginglistdisamb.R
\name{plot.TaggingListDisamb}
\alias{plot.TaggingListDisamb}
\title{Plot an object of type TaggingListDisamb}
\usage{
\method{plot}{TaggingListDisamb}(x, strict = TRUE, maxRank = NA, ...)
}
\arguments{
\item{show}{either "t" to show the threshold or "f" to show the f measure}
}
\description{
Plot an object of type TaggingListDisamb
}

