% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/tagginglistdisambbest.R
\name{plot.TaggingListDisambBest}
\alias{plot.TaggingListDisambBest}
\title{Plot an object of type TaggingListDisambBest}
\usage{
\method{plot}{TaggingListDisambBest}(x, strict = TRUE,
  addstripchart = FALSE, ...)
}
\arguments{
\item{x}{the object to plot}

\item{strict}{if TRUE plot the strict measures, otherwise the lenient measures}

\item{addstripchart}{if TRUE overlay the box plot with a strip chart}
}
\description{
This will plot a box plot of all the per-document F1 measures.
}

