% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/taggingscore.R
\name{plot.TaggingScore}
\alias{plot.TaggingScore}
\title{Plot an object of type TaggingScore}
\usage{
\method{plot}{TaggingScore}(x, strict = TRUE, show = "t", ...)
}
\arguments{
\item{show}{either "t" to show the threshold or "f" to show the f measure}
}
\description{
Plot an object of type TaggingScore
}

