% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/taggingbasic.R
\name{print.TaggingBasic}
\alias{print.TaggingBasic}
\title{Print a TaggingBasic object.}
\usage{
\method{print}{TaggingBasic}(x, ...)
}
\arguments{
\item{x}{the object to print}

\item{...}{additional parameters.}
}
\description{
Print a TaggingBasic object.
}

