% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/taggingdetaillistmatches.R
\name{print.TaggingDetailListMatches}
\alias{print.TaggingDetailListMatches}
\title{Print a TaggingDetailListMatches object.}
\usage{
\method{print}{TaggingDetailListMatches}(x, ...)
}
\arguments{
\item{x}{the object to print}

\item{...}{additional parameters.}
}
\description{
Print a TaggingDetailListMatches object.
}

