% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/tagginglistbest.R
\name{print.TaggingListBest}
\alias{print.TaggingListBest}
\title{Print a TaggingListBest object.}
\usage{
\method{print}{TaggingListBest}(x, ...)
}
\arguments{
\item{x}{the object to print}

\item{...}{additional parameters.}
}
\description{
Print a TaggingListBest object.
}

