% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/tagginglistdisambbest.R
\name{print.TaggingListDisambBest}
\alias{print.TaggingListDisambBest}
\title{Print a TaggingListDisambBest object.}
\usage{
\method{print}{TaggingListDisambBest}(x, ...)
}
\arguments{
\item{x}{the object to print}

\item{...}{additional parameters.}
}
\description{
Print a TaggingListDisambBest object.
}

