% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/tagginglistscore.R
\name{print.TaggingListScore}
\alias{print.TaggingListScore}
\title{Print a TaggingListScore object.}
\usage{
\method{print}{TaggingListScore}(x, ...)
}
\arguments{
\item{x}{the object to print}

\item{...}{additional parameters.}
}
\description{
Print a TaggingListScore object.
}

