% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/taggingscore.R
\name{print.TaggingScore}
\alias{print.TaggingScore}
\title{Print a TaggingScore object.}
\usage{
\method{print}{TaggingScore}(x, ...)
}
\arguments{
\item{x}{the object to print}

\item{...}{additional parameters.}
}
\description{
Print a TaggingScore object.
}

