% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/load.R
\name{select.GateEval}
\alias{select.GateEval}
\title{Select a specific evaluation id and type from the initial evaluation object.}
\usage{
\method{select}{GateEval}(x, evaluationId = NULL, evaluationType = NULL,
  annotationType = NULL)
}
\arguments{
\item{x}{The object created by GateEval}

\item{evaluationId}{A string that identifies the desired evaluatiob Id}

\item{evaluationType}{A string that identifies the desired evaluation type}

\item{annotationType}{A string the identifies the annotation type}
}
\value{
An object of some subclass of GateEval, depending on the kind of
selected evaluation instance.
}
\description{
Either evaluationId or evaluationType or both must be specified. If there
is no data for the given id and/or type, the method shows a warning and
returns NULL. Otherwise the method returns an object of a class specific
to the evaluation type.
}
\details{
NOTE: at the moment the data after performing the select must be restricted
to a single evaluationId, a single evaluationType and a single
annotationType (the special symbold "*" can be used to restrict the annotation
type to "all")

TODO: restricting to annotation type not yet implemented!!
}

