% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/load.R
\name{select}
\alias{select}
\title{Select one specific evaluation.}
\usage{
select(x, ...)
}
\arguments{
\item{x}{an existing evaluation object containing one or more evaluations.}

\item{...}{any additional parameters}
}
\value{
an object representing the selected evaluation instance
}
\description{
This tries to limit the data for an evaluation to a single evaluation id,
evaluation type, and annotation type. This is a necessary step before
specific kinds of analyses can be performed. Depending on what is selected,
objects of different sub-types of GateEval are returned.
It is an error to perform a selection which leaves nothing or more than
one specific evalaution.
}

